Atualizacao e correcao de erros -Criacao do endPoint para salvar endereco e buscar por ID -Alteracao do tipo do atributo dos models -Correcao de nome na Classe ProfissaoController

+++ src/main/java/com/pessoaDeon/domain/repository/PessoaRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
+public interface PessoaRepository extends JpaRepository<Pessoa, Integer> {
     @Override
-    Optional<Pessoa> findById(Long idPessoa);
+    Optional<Pessoa> findById(Integer idPessoa);
 
     Optional<Pessoa> findByCpf(String cpf);
